Tweaked the random InputStreat to take a long for the length.

+++ src/main/java/com/github/davidcarboni/cryptolite/Random.java
@@ -100,7 +100,7 @@ public class Random {
      * @param length The length of the stream.
      * @return {@link SecureRandom#nextBytes(byte[])}
      */
-    public static InputStream inputStream(final int length) {
+    public static InputStream inputStream(final long length) {
         return new InputStream() {
             int count;
             @Override
